Move Dates tests to new file
