+++ test/bookings-test.js
@@ -6,11 +6,6 @@ import {
   getTotalCost, 
   searchByDate, 
   searchByRoomType,
-  convertOneToTwoDigits,
-  configLongDate,
-  configShortDate,
-  getTodaysDate,
-  sortByDate
 } from '../src/bookings'
 
 describe('Bookings per customer', () => {
@@ -89,55 +84,5 @@ describe('Search rooms', () => {
     const xmasRooms = searchByDate(bookings, rooms, '2025/12/25')
     const onlyJuniorSuites = searchByRoomType(xmasRooms, 'junior suite');
     expect(onlyJuniorSuites).to.deep.equal(`We\'re terribly sorry - there are no rooms of that type available for the date you have selected. Please select a different room type, or book a different date.`);
-  });
-});
-
-describe('Dates', () => {
-  it('Should return a 2-digit number, given a 1-digit number', () => {
-    const month = 6;
-    const convertedMonth = convertOneToTwoDigits(month);
-    expect(convertedMonth).to.equal('06');
-  });
-
-  it('Should return the same number, given a 2-digit number', () => {
-    const day = 29;
-    const convertedDay = convertOneToTwoDigits(day);
-    expect(convertedDay).to.equal(29);
-  });
-
-  it('Should return the same number, given a number with more than 2 digits', () => {
-    const year = 2023;
-    const convertedYear = convertOneToTwoDigits(year);
-    expect(convertedYear).to.equal(2023);
-  });
-  
-  it('Should return a long date in the short format', () => {
-    const longDate = 'June 29, 2023';
-    const dataDate = configLongDate(longDate);
-    expect(dataDate).to.equal('2023/06/29');
-  });
-
-  it('Should return a different long date in the short format', () => {
-    const longDate = 'September 7, 2023';
-    const dataDate = configLongDate(longDate);
-    expect(dataDate).to.equal('2023/09/07');
-  });
-
-  it.only('Should return a short date in the long format', () => {
-    const shortDate = '2023/06/29';
-    const displayDate = configShortDate(shortDate);
-    expect(displayDate).to.equal('June 29, 2023')
-  });
-
-  it.only('Should return a different short date in the long format', () => {
-    const shortDate = '2023/09/07';
-    const displayDate = configShortDate(shortDate);
-    expect(displayDate).to.equal('September 7, 2023')
-  });
-
-  it('Should sort a user\'s bookings in reverse chronological order', () => {
-    const leathasBookings = sampleData.customers[0].bookings;
-    const sortedBookings = sortByDate(leathasBookings);
-    expect(sortedBookings).to.deep.equal(sampleData.leathasBookingsSorted);
   });
 });
